Retrieve records from dblp API

Retrieve records from dblp API

dblp_search(
  query = NULL,
  start = 1,
  n = 10,
  c = NULL,
  api_config = NULL,
  entity = c("publications", "authors", "venues")
)

Arguments

query
the query string to search for, as described in https://dblp.org/faq/13501473.html, for example q=test+search

start
the "f"irst hit in the numbered sequence of search results to return. In combination with the h parameter, this parameter can be used for pagination of search results

n
maximum number of search results ("h"its) to return. For bandwidth reasons, this number is capped at 1000

c
maximum number of completion terms to return. For bandwidth reasons, this number is capped at 1000

api_config
a configuration setting for the API including base URL etc, by default from config()

entity
one of "publications", "authors" or "venues" where "publications" is default

Value

results records returned from the search

Examples

if (FALSE) {
dblp_search()
}